+++ docs/src/examples/QasGalleryCard/PdfAutomaticFileDetection.vue
@@ -9,11 +9,10 @@
 </template>
 
 <script setup>
-defineOptions({ name: 'PdfFile' })
+defineOptions({ name: 'PdfAutomaticFileDetection' })
 
 const galleryCardProps = {
   useHeaderErrorMode: false,
-  fileType: 'pdf',
-  url: 'https://dhaw-crmap01-assets.s3.amazonaws.com/uploads/customers/import/upload/43e8aaef-6237-4e51-9cfd-f6b2c923f152.pdf'
+  url: 'https://dhaw-crmap01-assets.s3.amazonaws.com/uploads/customers/import/upload/43e8aaef-6237-4e51-9cfd-f6b2c923f152.pdf' // detecção automática do fileType pela URL
 }
 </script>